Yangilanish Loyhada

+++ src/users/users.service.ts
@@ -1,4 +1,3 @@
-import { ImageService } from './../image/image.service';
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
@@ -6,52 +5,66 @@ import { User } from './scheam/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import * as bcrypt from "bcrypt";
-import { NotFoundError } from 'rxjs';
-import { ref } from 'process';
 import { JwtService } from '@nestjs/jwt';
 import { MailService } from '../mail/mail.service';
-import { v4 as uuid } from 'uuid';
-
+import { v4 as uuidv4 } from 'uuid';
+import { Express } from 'express';
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User) private readonly userRepo: Repository<User>,
-    private readonly imageService:ImageService,
     private readonly jwtService: JwtService,
     private readonly mailService:MailService
 
   ) {}
-  async create(createUserDto: CreateUserDto) {
-    const { email, password, confrim_password , imageId} = createUserDto;
-    const images = await this.imageService.findOne(imageId)
-    if(!images){
-      throw new BadRequestException({message:"Bunday Id Malumot Yoq"})
-    }
-    console.log(images);
+  async create(createUserDto: CreateUserDto, image: Express.Multer.File) {
+    const { email, password, confrim_password , phone_number} = createUserDto;
     if (password !== confrim_password) {
       console.log(password);
       console.log(confrim_password);
       throw new BadRequestException("Parollar mos emas");
     }
-    const existingAdmin = await this.userRepo.findOne({ where: { email } });
+    // const existingAdmin = await this.userRepo.findOne({ where: { email } });
+    // if (existingAdmin) {
+    //   throw new BadRequestException("Bunday emailli foydalanuvchi mavjud");
+    // }
+    const existingAdmin = await this.userRepo.findOne({ where: [{ phone_number }] });
     if (existingAdmin) {
-      throw new BadRequestException("Bunday emailli foydalanuvchi mavjud");
+      throw new BadRequestException("Bunday telefon raqamli foydalanuvchi mavjud");
     }
-    const hashed_password = await bcrypt.hash(password, 7);
-    const newUser = this.userRepo.create({ ...createUserDto, hashed_password:hashed_password,image:images});
+    const password_hashed = await bcrypt.hash(password, 7);
+    console.log("Parol:", password); // Murod1972!
+    console.log("Hash:", password_hashed);
+    const activation_link = uuidv4();
+    const imagePath = image ? image.path : "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_960_720.png";
+
+// Alohida o‘rnatib chiqamiz:
+const newUser = this.userRepo.create({
+  first_name: createUserDto.first_name,
+  last_name: createUserDto.last_name,
+  email: createUserDto.email,
+  phone_number: createUserDto.phone_number,
+  hashed_password:password_hashed,
+  activation_link,
+  image: imagePath,
+  is_active:true
+});
+    console.log("User Servisdagi User=>",newUser);
     try {
       this.mailService.sendMailUser(newUser)
     } catch (error) {
       console.log("Emailga Hat Yuborishda Hatolik Yuzaga Keldi");
     }
-    return await this.userRepo.save(newUser);
+    const a =  await this.userRepo.save(newUser);
+    console.log("Bazaga saqlangan foydalanuvchi:", a.hashed_password);
+    return a
   }
   findAll() {
-    return this.userRepo.find({relations:["image"]});
+    return this.userRepo.find();
   }
 
   async findOne(id: number) {
-    const user = await this.userRepo.findOne({ where: { id } ,relations:["image"]});
+    const user = await this.userRepo.findOne({ where: { id }});
     if(!user){
       throw new BadRequestException({message:"Bunday Id Malumot Yoq "})
     }
@@ -61,12 +74,6 @@ export class UsersService {
     return this.userRepo.findOne({ where: { email } });
   }
   async update(id: number, updateUserDto: UpdateUserDto) {
-    if(updateUserDto.imageId){
-      const images = await this.imageService.findOne(updateUserDto.imageId)
-      if(!images){
-        throw new BadRequestException({message:'Bunday Image Id Mavjud emas'})
-      }
-    }
     const user = await this.userRepo.findOne({where:{id}})
     if(!user){
       throw new BadRequestException({message:"Bunday Id Malumotlari mavjud emas"})
@@ -84,14 +91,19 @@ export class UsersService {
     return this.userRepo.update(id, updateUserDto);  
   }
   remove(id: number) {
+    const user = this.userRepo.findOne({where:{id}})
+    if(!user){
+      throw new BadRequestException({message:"Bunday Id Malumotlari mavjud emas"})
+    }
+
     return this.userRepo.delete(id);
   }
-  async findByToken(refreshToken:string){
-    if(refreshToken){
+  async findByToken(refresh_token:string){
+    if(!refresh_token){
       throw new NotFoundException({massage:"Bunday token mavjud emas"})
     }
     try {
-      const decoded = this.jwtService.verify(refreshToken, {
+      const decoded = this.jwtService.verify(refresh_token, {
         secret:process.env.REFRESH_TOKEN_KEY
       }) as {id:number}
       const admin = await this.findOne(decoded.id)
@@ -129,4 +141,11 @@ export class UsersService {
   async save(user:User) {
     return this.userRepo.save(user);
   }
+  
+  async updateRefreshToken(id: number, hashed_refresh_token: string | null): Promise<void> {
+    const result = await this.userRepo.update(id, { hashed_refresh_token });
+      if (result.affected === 0) {
+        throw new NotFoundException(`User with ID ${id} not found`);
+      }
+    }
 }
